Move old Builder tests to query_test module
